feat(schedule): 스케줄 관리 기능 구현 - 2
- 공통, 기본 근무 일정, 실제 근무 일정으로 디렉토리 구조 변경
- 근무 일정 엔티티 데이터 구조 변경
- CRUD에 모두 DTO 적용

Issue: #1
