+++ src/main/java/com/ss6051/backendspring/schedule/actual/dto/ActualWorkCreationDTO.java
@@ -1,4 +1,4 @@
-package com.ss6051.backendspring.schedule.dto.actual;
+package com.ss6051.backendspring.schedule.actual.dto;
 
 /**
  * 실제 근무 시간을 생성하기 위한 DTO
