+++ src/main/java/com/ss6051/backendspring/schedule/basic/dto/BasicWorkUpdateDTO.java
@@ -1,13 +1,13 @@
-package com.ss6051.backendspring.schedule.dto.basic;
+package com.ss6051.backendspring.schedule.basic.dto;
 
 /**
  * 기본 근무 시간을 생성하기 위한 DTO
  *
- * @param storeId      매장 ID
- * @param accountId    계정 ID
+ * @param storeId   매장 ID
+ * @param accountId 계정 ID
  * @param basicWorkDTO 기본 근무 시간 DTO
  */
-public record BasicWorkCreationDTO(
+public record BasicWorkUpdateDTO(
         long storeId,
         long accountId,
         BasicWorkDTO basicWorkDTO) {
