+++ src/main/java/com/ss6051/backendspring/schedule/common/ScheduleRepository.java
@@ -1,6 +1,6 @@
-package com.ss6051.backendspring.schedule.repository;
+package com.ss6051.backendspring.schedule.common;
 
-import com.ss6051.backendspring.schedule.domain.Schedule;
+import com.ss6051.backendspring.schedule.common.domain.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
