+++ src/main/java/com/ss6051/backendspring/schedule/common/domain/ScheduleAccountPair.java
@@ -1,0 +1,6 @@
+package com.ss6051.backendspring.schedule.common.domain;
+
+import com.ss6051.backendspring.global.domain.Account;
+
+public record ScheduleAccountPair(Schedule schedule, Account account) {
+}
